refactor: 디렉토리 구조 layout, section으로 분리
