+++ src/app/_component/layout/Header.tsx
@@ -1,0 +1,6 @@
+import { PropsWithChildren } from 'react';
+import * as styles from './header.css';
+
+export default function Header({ children }: PropsWithChildren) {
+  return <div className={styles.navigationHeader}>{children}</div>;
+}
