+++ src/app/page.tsx
@@ -1,11 +1,14 @@
-import NavigationHeader from './_component/NavigationHeader';
-import Main from './_component/Main';
-import Footer from './_component/Footer';
+import Header from '@app/_component/layout/Header';
+import LogoHeader from '@app/_component/LogoHeader';
+import Main from '@app/_component/layout/Main';
+import Footer from '@app/_component/layout/Footer';
 
 export default function Home() {
   return (
     <>
-      <NavigationHeader />
+      <Header>
+        <LogoHeader />
+      </Header>
       <Main />
       <Footer />
     </>
