+++ src/lib/const/carousel.ts
@@ -1,16 +1,14 @@
-import { Carousel } from '@lib/types';
-
-export const CAROUSEL: Carousel = {
-  1: {
+export const CAROUSEL = [
+  {
     mobile: '/carousel/carousel_mobile_1.png',
     desktop: '/carousel/carousel_1.png',
   },
-  2: {
+  {
     mobile: '/carousel/carousel_mobile_2.png',
     desktop: '/carousel/carousel_2.png',
   },
-  3: {
+  {
     mobile: '/carousel/carousel_mobile_3.png',
     desktop: '/carousel/carousel_3.png',
   },
-};
+] as const;
