--- src/lib/types/carousel.type.ts
@@ -1,6 +1,0 @@
-export type Carousel = {
-  [key: number]: {
-    mobile: string;
-    desktop: string;
-  };
-};
