+++ src/lib/types/index.ts
@@ -1,3 +1,2 @@
-export * from './carousel.type';
 export * from './featureInfo.type';
 export * from './pageAddress.type';
